requirements.txt
auto_mix_prep==0.2.0
Pillow==11.1.0
datasets==3.0.1
evaluate==0.4.3
ipdb==0.13.13
numpy==1.26.4
pandas==2.2.3
Requests==2.32.3
scikit_learn==1.6.1
bitsandbytes==0.45.2
torch==2.5.1
tqdm==4.66.4
transformers==4.49.0
sentence_transformers==3.3.1
torchvision==0.20.1
vllm==0.6.4.post1
pydantic==2.10.6
pydantic_settings==2.8.1
uvicorn==0.34.0
python-settings
fastapi==0.115.8
faiss-gpu-cu12== 1.10.0
python-multipart==0.0.20
qwen-vl-utils==0.0.10
accelerate==1.2.1
nervaluate==0.2.0
SPARQLWrapper==2.0.0
haversine==2.9.0
requests-cache==1.2.1
openai==1.55.0
websockets==13.1
trl==0.13.0
rq==2.3.1
redis==5.2.1
peft==0.13.2
mastodon_py==2.0.1
huggingface_hub==0.30.2
geopy==2.4.1
fuzzywuzzy==0.18.0
celery==5.4.0
beautifulsoup4==4.13.3
httpx==0.27.2
